requirements.txt
opencv_contrib_python==4.7.0.72
Pillow==9.5.0
PyYAML==6.0
torch==1.13.0
torchvision==0.14.0
